batch: Move selected-line source refresh
The batch source refresh path repairs stale selections by advancing
sources, remapping ownership, and re-annotating selected line entries.

Line re-annotation still lives in the orchestration module beside session
cache updates and ownership merge preparation. That makes callers depend on
a larger module when they only need selected-line coordinate repair.

This commit addresses that split by moving selected-line coordinate refresh
into selected_line_source_refresh. Source refresh and direct callers now
import that narrower API without compatibility aliases.

A following tests commit will update import-boundary coverage and focused
tests for the moved selected-line API.
